Optimize Merkle root verification for blocks [ECR-1622] (#802)
* Optimize Merkle root verification for blocks [ECR-1622]

PR #664 refactored block verification but it uses ProofListIndex to
compute Merkle root hash. As suggested during code review, we can
replace ProofListIndex and MemoryDB with direct computation of the root
hash.

The tests cross-verify this implementation with ProofListIndex to ensure
that they are consistent. However, I'm not sure whether we need this
cross-verification (only checking hash values may be enough).
Furthermore, maybe merkle_root() function can be moved to some more
appropriate module.

* Move Merkle root computation to a helper module

Introduce a new helper module "merkle" which will contain utilities
for dealing with Merkle trees. Maybe some day we will use this as
common code with ProofListIndex.

The functions are made public so add proper documentation comments
for them.

* Make root_hash() generic over iterators

As this is a public function we should not limit it to just slices
of Hashes. Generalize it over any iterator over references to Hashes
which we will need to copy them into a temporary Vec.

* Improve documentation comment

* rustfmt again

* Revert "Make root_hash() generic over iterators"

This reverts commit 640f6e07e730d06d291778024fe3d6ef7f745b37.

Minor conflicts in doc comments, nothing serious.

* Compute root hashes in-place

Avoid allocating intermediate Vecs by computing the hashes in-place.
The logic is a bit convoluted as we have to account for proper indexing,
but this should be more efficient.

We still do a full slice copy to get the initial Vec, and we may need to
allocate an additional empty hash if the optimizer is not that smart.

* Move hash_{one,pair} functions to proof_list_index module

Let's keep hash combination functions in one place. This is important
as ProofListIndex uses slightly special combination for the case when
there is only one hash (i.e., when there is odd number of hashes to
combine): it computes hash(h) instead of arguably more canonical
hash_pair(h, h).

Leave the functions crate-private as we need them in helpers::merkle,
do not export them publicly.

* Test root_hash() only against ProofListIndex

Do not expose concrete Hash values in tests, just verify that Hash
instaces are equal.

* Move root_hash() to proof_list_index module

It makes more sense to keep this utility function close to
ProofListIndex. It is effectively an optimized way to compute the same
root hash as available via ProofListIndex.

Remove now unnecessary helpers::merkle module. All Merkle-related code
is now located in the "storage" module.

* Update CHANGELOG

* Revert "Compute root hashes in-place"

This reverts commit 3e60b21e6a7d79ec1af9c400d79ccc1e501ea358.

After all, I did not really like the optimized in-place implementation.
It has too much magic and weird C-style indexing in it which is not
idiomatic for Rust. I'd guess we'll keep the original implementation
until profiling will show that this is really a bottleneck and we need
to do something about it.

* Avoid unnecessary initial Vec allocation

Strictly speaking, we can avoid cloning the whole original Vec if we use
combine_hash_list() to allocate the first Vec with half of the original
hashes.

However, this also requires us to explicitly handle the special case of
hash list of length 1 (we need to simply return the hash instead of
rehashing it).

* Apply rustfmt

* Update CHANGELOG formatting and grammar

* Use destructuring matching for pairs

More explicit and arguably more readable. Plus we handle the case
when the 'pair' has unexpected argument count.